Remove expectations on secs_per_epoch metric.
Technically the tests can be run on any platform and the expectations might change resulting in tests failing. We can introduce such assertions when we integrate the benchmarks with a CI or Regression analyzer.

We do export these metrics, so still the external integration without any change should be possible.

PiperOrigin-RevId: 321594202