Add automatic named call wrapping for profiling.
Because named call wraps every method in a model, out of paranoia
we keep it off by default to lower python overheads until we've
measured the cost of having it on, but one can enable this mode by
calling nn.enable_named_call() and later disable it with
nn.disable_named_call().  The wrapping occurs at subclass creation
time.