Add FrozenKeysView and FrozenValuesView to flax.
Currently, the `flax.core.FrozenDict` class get the `.keys()` method as a mix-in from subclassing `Mapping`. This results in the `.keys()` method return the default `collections.abc.KeysVeiw` object. The `__repr__` for this default keys view is unhelpful, as it essentially just prints the `__repr__` of the frozen dict itself. This becomes annoying when the ForzenDict is used for the parameters of a large network, as the `__repr__` become very large.

This adds a new `FrozenKeysView` subclass for `collections.abc.KeysView` that provides a new `__repr__` so that the output is similar to that of `.keys()` for a normal python dictionary or running `list(my_forzen_dict.keys())`.

I would assume that these subclasses would inherit the view-like behavior where changes in the underlying dict are reflected in the keys view, but because the frozen dict is immutable, and any changes are therefore different objects, we don't have worry about this behavior.

It also creates an analogouos `FrozenValuesView` and updates the frozen dicts `.keys()` and `.values()` methods to return these new subclasses.

PiperOrigin-RevId: 386026965