Forbid passing an int as kernel_size
Users expect that the int will be broadcasted to the appropiate rank.
But this is ambigious because we don't know if
or how many batch dimensions there are.

In contract, for stride and dilation we can broadcast
because the kernel rank is already known. This PR
therefore implements int broadcasting for these args.