Add `unroll` to `jax_utils.scan_in_dim`
The `unroll` flag within `jax.lax.scan` allows one to unroll a specified
number of `scan` iterations to unroll into a single iteration of the loop.

Currently, this change allows for `unroll` to be a tuple of positive
integers, allowing for one to control the number of unrolls which can
occur with each dimension being looped over.