Fixes a bug in DenseGeneral.
Prior to this PR, it was not possible to specify the `dtype` of the initializer functions for `DenseGeneral`, because the wrapper function takes a dtype argument, but one cannot call this wrapper function with a custom dtype argument. The solution in this PR is to disregard setting dtypes in the wrapper all-together, and leave it up to the user to specify the dtype in the init functions. I've also added a test to demonstrate this behavior.

PiperOrigin-RevId: 416027340